pyproject.toml
[project]
name = "table-dataset-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "azure-storage-blob==12.27.0",
    "dotenv>=0.9.9",
    "fastapi>=0.120.0",
    "gradio>=5.49.1",
    "joblib>=1.5.2",
    "nest-asyncio>=1.6.0",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymupdf>=1.26.5",
    "requests>=2.32.5",
    "theflow>=0.8.6",
    "uvicorn>=0.38.0",
]
